+++ test/function.ts
@@ -13,7 +13,7 @@ import {
   unsafeCoerce,
   absurd,
   flow,
-  unary
+  tupled
 } from '../src/function'
 
 const f = (n: number) => n + 1
@@ -80,11 +80,11 @@ describe('function', () => {
     assert.deepStrictEqual(flow(f, g, f, g, f, g, f, g, f)(2), 63)
   })
 
-  it('unary', () => {
+  it('tupled', () => {
     const f1 = (a: number): number => a * 2
     const f2 = (a: number, b: number): number => a + b
-    const u1 = unary(f1)
-    const u2 = unary(f2)
+    const u1 = tupled(f1)
+    const u2 = tupled(f2)
     assert.deepStrictEqual(u1.length, 1)
     assert.deepStrictEqual(u1([1]), 2)
     assert.deepStrictEqual(u2.length, 1)